chore: add new .htaccess file in domain root directory
This file configures the apache server from a share-hosting as laravel
index.php file is not located in the root main directory but in a
subdirectory called "/public". with this configuration file the index.php
from laravel project becomes accessible when the domain url is request
and then the entire project would run without any problem.

Authored-by: Jhonatan Martinez