feat: add partial view for website navbar
This navbar.blade.php file can be use in all website pages. this navbar
just have one static option, this is the index page. the other options
can be add dinamically through a module in the website administrator
panel. you can add two types of options to the the navigation bar, links
and dropdowns.

Authored-by: Jhonatan Martinez