refactor: remove CSS property from navbar
the navigation bar had a property "position" with a value of "absolute"
that is applied when the screen min-width is 992px.

the "position: absolute" CSS property was used to place the navigation bar
in top of a carousel, as the carousel was used as background in the hero
section.

since the carousel is not used any more as background for the
navigation bar and it is now place outside of the hero section, there was
blank DIV remaining.

this property was remove to avoid the creation of a blank space because
of the extra DIV that came with the template

Authored-by: Jhonatan Martinez